Add LearnableChannel and SpeakableChannel
Channel types that Parrot is able to learn in and speak in, respectively. Felt the need to make this abstraction when I realized that TextChannel isn't hte only type Parrot can speak in, and we were artibtrarily procluding it from being able to speak in, say, threads, because we were checking it to be just a TextChannel when it could have gotten away with it being something else.

Comes with respective type guard functions to replace the raw `isinstance` calls that peppered the codebase.

This arguably makes the code's meaning more clear, now that there is a distinction between checking a channel's type for whether we can learn in it, and whether we can speak in it.